Rename `serve.ts` to `server.ts`
